fix: 修正未登錄移動到 add-card 時的轉跳 refactor: 重構 add-card 的 api 呼叫
