+++ components/features/landing-page/Guide/index.jsx
@@ -1,8 +1,16 @@
 import Link from 'next/link';
 import Image from 'next/image';
+import { useEffect, useState } from 'react';
 import Button from '../../../common/Button/Button';
 
 function Guide() {
+  const [isLogin, setIsLogin] = useState(false);
+
+  useEffect(() => {
+    const loginStatus = localStorage.getItem('auth');
+    if (loginStatus?.startsWith('Bear')) setIsLogin(true);
+  }, []);
+
   return (
     <div className="flex flex-col items-center bg-main-02 px-5 pt-11 pb-12 font-bold laptop:pt-12 laptop:pb-16">
       <Image
@@ -17,7 +25,7 @@ function Guide() {
         <br className="laptop:hidden" />
         幫你的名片施加魔法🪄
       </h3>
-      <Link href="/add-card">
+      <Link href={isLogin ? '/add-card' : '/login'}>
         <Button
           variant="contained"
           className="w-52 bg-main-04 duration-200 hover:bg-danger"
